Start moving Schema Dialect features to Schema/Dialect
Move code + tests for converting columns/indexes.
