+++ lib/Cake/Database/SqlDialectTrait.php
@@ -176,14 +176,4 @@ trait SqlDialectTrait {
 		return 'ROLLBACK TO SAVEPOINT LEVEL' . $name;
 	}
 
-/**
- * Get extra schema metadata columns
- *
- * This method returns information about additional metadata present in the data
- * generated by describeTableSql
- *
- * @return void
- */
-	abstract function extraSchemaColumns();
-
 }
