+++ lib/Cake/Test/TestCase/Database/Driver/MysqlTest.php
@@ -139,110 +139,6 @@ SQL;
 	}
 
 /**
- * Dataprovider for column testing
- *
- * @return array
- */
-	public static function columnProvider() {
-		return [
-			[
-				'DATETIME',
-				['datetime', null]
-			],
-			[
-				'DATE',
-				['date', null]
-			],
-			[
-				'TIME',
-				['time', null]
-			],
-			[
-				'TINYINT(1)',
-				['boolean', null]
-			],
-			[
-				'TINYINT(2)',
-				['integer', 2]
-			],
-			[
-				'INTEGER(11)',
-				['integer', 11]
-			],
-			[
-				'BIGINT',
-				['biginteger', null]
-			],
-			[
-				'VARCHAR(255)',
-				['string', 255]
-			],
-			[
-				'CHAR(25)',
-				['string', 25]
-			],
-			[
-				'TINYTEXT',
-				['string', null]
-			],
-			[
-				'BLOB',
-				['binary', null]
-			],
-			[
-				'MEDIUMBLOB',
-				['binary', null]
-			],
-			[
-				'FLOAT',
-				['float', null]
-			],
-			[
-				'DOUBLE',
-				['float', null]
-			],
-			[
-				'DECIMAL(11,2)',
-				['decimal', null]
-			],
-		];
-	}
-
-/**
- * Test parsing MySQL column types.
- *
- * @dataProvider columnProvider
- * @return void
- */
-	public function testConvertColumnType($input, $expected) {
-		$driver = new Mysql();
-		$this->assertEquals($expected, $driver->convertColumn($input));
-	}
-
-/**
- * Provider for testing index conversion
- *
- * @return array
- */
-	public static function convertIndexProvider() {
-		return [
-			['PRI', 'primary'],
-			['UNI', 'unique'],
-			['MUL', 'index'],
-		];
-	}
-/**
- * Test parsing MySQL index types.
- *
- * @dataProvider convertIndexProvider
- * @return void
- */
-	public function testConvertIndex($input, $expected) {
-		$driver = new Mysql();
-		$this->assertEquals($expected, $driver->convertIndex($input));
-	}
-
-/**
  * Test listing tables with Mysql
  *
  * @return void
